# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "IMOK Api",
        "description": "This API is meant to be used along with the IMOK_Mobile and IMOK_Desktop apps. All routes require you to be authenticated, so your first step is to send your credentials to api/auth/login and get a bearer token.",
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "http://imok.am.manusien-ecolelamanu.fr/imok_api/current/public/api",
            "description": "API BASE PATH"
        }
    ],
    "paths": {
...
        "/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login",
                "operationId": "App\\Http\\Controllers\\AuthController::login",
                "requestBody": {
                    "description": "Each field is required",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "LOGIN",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "LOGIN",
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Logged in"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout",
                "operationId": "App\\Http\\Controllers\\AuthController::logout",
                "responses": {
                    "200": {
                        "description": "Successfully logged out"
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
        "/auth/refresh": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Refresh token",
                "operationId": "App\\Http\\Controllers\\AuthController::refresh",
                "responses": {
                    "200": {
                        "description": "Token refreshed"
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
        "/auth/me": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Who am I",
                "operationId": "App\\Http\\Controllers\\AuthController::me",
                "responses": {
                    "200": {
                        "description": "User"
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "description": "JWT Auth Token",
                "name": "JWT",
                "in": "header"
            }
        }
    }
}